chore: Organize file locations
